+++ src/main/java/shop/tryit/repository/order/OrderCustom.java
@@ -3,7 +3,7 @@ package shop.tryit.repository.order;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import shop.tryit.domain.member.Member;
-import shop.tryit.domain.order.OrderSearchDto;
+import shop.tryit.domain.order.dto.OrderSearchDto;
 
 public interface OrderCustom {
