+++ src/main/java/shop/tryit/repository/order/OrderDetailCustom.java
@@ -2,7 +2,7 @@ package shop.tryit.repository.order;
 
 import java.util.List;
 import shop.tryit.domain.order.Order;
-import shop.tryit.domain.order.OrderDetailSearchDto;
+import shop.tryit.domain.order.dto.OrderDetailSearchDto;
 
 public interface OrderDetailCustom {
